Fix: free_netdev called twice
This device defines a destructor to `free_netdev`, which conflict with
the module's exit code (a device will be freed twice)

Not sure if this is a good fix. What if a kernel module can generate
serveral devices like the `tun` module? Should we free the device using
destructor, or using module's exit function?